media-libs/lvtk: depend on graphviz if doc is to be built, do not move documentation if it is not built

+++ overlays/proaudio/media-libs/lvtk/lvtk-1.1.0.ebuild
@@ -22,6 +22,10 @@ RDEPEND="media-libs/lv2
 	>=dev-libs/boost-1.40.0"
 DEPEND="${RDEPEND}
 	${PYTHON_DEPS}
+	doc? (
+		app-doc/doxygen
+		media-gfx/graphviz
+	)
 	virtual/pkgconfig"
 
 DOCS=( AUTHORS ChangeLog README )
@@ -41,7 +45,10 @@ src_configure() {
 
 src_install() {
 	waf-utils_src_install
+
 	# It does not respect docdir properly, reported upstream
-	mv "${ED}/usr/share/doc/${PF}/lvtk-1.0/html" "${ED}/usr/share/doc/${PF}/html"
-	rmdir "${ED}/usr/share/doc/${PF}/lvtk-1.0"
+	if use doc; then
+		mv "${ED}/usr/share/doc/${PF}/lvtk-1.0/html" "${ED}/usr/share/doc/${PF}/html"
+		rmdir "${ED}/usr/share/doc/${PF}/lvtk-1.0"
+	fi
 }
